LPD-17663 Moves tests to AssetCategoryLocalServiceTest and removes class
